moved logger into util package
